fixing 1

+++ resources/views/admin/index.blade.php
@@ -43,7 +43,6 @@
                   <th>No.</th>
                   <th>Nama</th>
                   <th>Email</th>
-                  <th>Kode Unik</th>
                   <th>Level</th>
                   <th colspan="2" class="text-center">Aksi</th>                        </tr>
                 </thead>
@@ -55,14 +54,13 @@
                   <th scope="row">{{$no++}}</th>
                   <td>{{$p->name}}</td>
                   <td>{{$p->email}}</td>
-                  <td>{{$p->nip}}</td>
                   <td>{{$p->admin}}</td>
 
                   <td class="mx-1 px-1">
                     <a href="/admin/edit/{{ $p->id }}" class="btn btn-sm btn-warning">Edit</a>
                   </td>
                   <td class="mx-1 px-0">
-                    <a href="/admin/hapus/{{ $p->name }}" class="btn btn-sm btn-danger">Hapus</a>
+                    <a href="/admin/hapus/{{ $p->name }}" class="btn btn-sm btn-danger" onclick="return confirm('Anda yakin ingin menghapus user ini ?')">Hapus</a>
                   </td>
                 </tr>
                 @endforeach
